Go code gen for arrays, union vectors, and with path-aware imports (#11)
This change adds a very basic implementation of support for arrays and
union vectors in Go.

It also implements go package naming using file path and namespace.

This package import schema allows for automated Bazel bulid rule
generation with Gazelle while also keeping the directory structure of
generated files aligned with the structure of the .fbs files.

This package naming method is fairly arbitrary. Ideally we would
introduce a "go_package" option, or use Bazel rule name as the package.

This has been only tested with --gen-one-file mode.

Each .fbs file gets a .go file. The package name of the go file is the
final component of the .fbs namespace.

The import path for the file is `dirname(.fbs) + /package_name`.

For example, for the following .fbs file, we would have the following import:

.fbs file: messages/foo/bar.fbs // defines namespace foo
import:    messages/foo/foo

.fbs file: messages/foo/bar/baz.fbs // defines namespace baz
import:    messages/foo/bar/baz